// src/api/generated/vegnett/types.gen.ts
export type Kryssystem = {
    kryssystem: number;
    kryssdel: number;
    retning: Retning;
    trafikantgruppe: Trafikantgruppe;
    meter?: number;
    fra_meter?: number;
    til_meter?: number;
};
// ...
export type Sideanlegg = {
    sideanlegg: number;
    sideanleggsdel: number;
    retning: Retning;
    trafikantgruppe: Trafikantgruppe;
    meter?: number;
    fra_meter?: number;
    til_meter?: number;
};

export type Sideposisjon = 'MH' | 'MV' | 'VT' | 'M' | 'H' | 'HT' | 'HV' | 'K' | 'V' | 'L' | 'R0';

export type SosiMedium = 'T' | 'B' | 'L' | 'U' | 'S' | 'O' | 'V' | 'D' | 'I' | 'W' | 'J' | 'X' | 'IKKE_REGISTRERT';

export type Strekning = {
    strekning: number;
    delstrekning: number;
    arm: boolean;
    adskilte_løp: AdskilteLop;
    trafikantgruppe: Trafikantgruppe;
    retning: Retning;
    meter?: number;
    fra_meter?: number;
    til_meter?: number;
    adskilte_løp_nummer?: string;
};
// ...
export type Vegsystem = {
    vegkategori: Vegkategori;
    fase: Vegfase;
    nummer?: number;
};

export type Vegsystemreferanse = {
    vegsystem?: Vegsystem;
    strekning?: Strekning;
    kryssystem?: Kryssystem;
    sideanlegg?: Sideanlegg;
    kortform?: string;
};
// ...
export type PosisjonGeometri = {
    wkt: string;
    srid: number;
};

export type PosisjonVeglenkesekvens = {
    veglenkesekvensid: number;
    relativPosisjon: number;
    kortform: string;
};

export type AlwaysIncludeMapStringPosisjon = {
    empty?: boolean;
    [key: string]: Posisjon | boolean | undefined;
};

export type PosisjonMedAvstand = {
    vegsystemreferanse: Vegsystemreferanse;
    veglenkesekvens: PosisjonVeglenkesekvens;
    geometri: PosisjonGeometri;
    kommune: number;
    avstand: number;
};